+++ tests/test_routers.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
+import warnings
 from collections import namedtuple
 
 import pytest
@@ -86,13 +87,13 @@ kwarged_notes_router = SimpleRouter()
 kwarged_notes_router.register(r'notes', KWargedNoteViewSet)
 
 namespaced_router = DefaultRouter()
-namespaced_router.register(r'example', MockViewSet, base_name='example')
+namespaced_router.register(r'example', MockViewSet, basename='example')
 
 empty_prefix_router = SimpleRouter()
-empty_prefix_router.register(r'', EmptyPrefixViewSet, base_name='empty_prefix')
+empty_prefix_router.register(r'', EmptyPrefixViewSet, basename='empty_prefix')
 
 regex_url_path_router = SimpleRouter()
-regex_url_path_router.register(r'', RegexUrlPathViewSet, base_name='regex')
+regex_url_path_router.register(r'', RegexUrlPathViewSet, basename='regex')
 
 
 class BasicViewSet(viewsets.ViewSet):
@@ -319,7 +320,7 @@ class TestActionKeywordArgs(TestCase):
                 })
 
         self.router = SimpleRouter()
-        self.router.register(r'test', TestViewSet, base_name='test')
+        self.router.register(r'test', TestViewSet, basename='test')
         self.view = self.router.urls[-1].callback
 
     def test_action_kwargs(self):
@@ -344,7 +345,7 @@ class TestActionAppliedToExistingRoute(TestCase):
                 })
 
         self.router = SimpleRouter()
-        self.router.register(r'test', TestViewSet, base_name='test')
+        self.router.register(r'test', TestViewSet, basename='test')
 
         with pytest.raises(ImproperlyConfigured):
             self.router.urls
@@ -481,3 +482,71 @@ class TestViewInitkwargs(URLPatternsTestCase, TestCase):
         initkwargs = match.func.initkwargs
 
         assert initkwargs['basename'] == 'routertestmodel'
+
+
+class TestBaseNameRename(TestCase):
+
+    def test_base_name_and_basename_assertion(self):
+        router = SimpleRouter()
+
+        msg = "Do not provide both the `basename` and `base_name` arguments."
+        with warnings.catch_warnings(record=True) as w, \
+                self.assertRaisesMessage(AssertionError, msg):
+            warnings.simplefilter('always')
+            router.register('mock', MockViewSet, 'mock', base_name='mock')
+
+        msg = "The `base_name` argument has been deprecated in favor of `basename`."
+        assert len(w) == 1
+        assert str(w[0].message) == msg
+
+    def test_base_name_argument_deprecation(self):
+        router = SimpleRouter()
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+            router.register('mock', MockViewSet, base_name='mock')
+
+        msg = "The `base_name` argument has been deprecated in favor of `basename`."
+        assert len(w) == 1
+        assert str(w[0].message) == msg
+        assert router.registry == [
+            ('mock', MockViewSet, 'mock'),
+        ]
+
+    def test_basename_argument_no_warnings(self):
+        router = SimpleRouter()
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+            router.register('mock', MockViewSet, basename='mock')
+
+        assert len(w) == 0
+        assert router.registry == [
+            ('mock', MockViewSet, 'mock'),
+        ]
+
+    def test_get_default_base_name_deprecation(self):
+        msg = "`CustomRouter.get_default_base_name` method should be renamed `get_default_basename`."
+
+        # Class definition should raise a warning
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+
+            class CustomRouter(SimpleRouter):
+                def get_default_base_name(self, viewset):
+                    return 'foo'
+
+        assert len(w) == 1
+        assert str(w[0].message) == msg
+
+        # Deprecated method implementation should still be called
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+
+            router = CustomRouter()
+            router.register('mock', MockViewSet)
+
+        assert len(w) == 0
+        assert router.registry == [
+            ('mock', MockViewSet, 'foo'),
+        ]